Move dependencyUpdate logic to CommonBuildPlugin.
