+++ core/build.gradle.kts
@@ -1,8 +1,9 @@
+import com.github.benmanes.gradle.versions.updates.DependencyUpdatesTask
+
 plugins {
     id("com.android.library")
     kotlin("android")
     kotlin("android.extensions")
-    id("com.github.ben-manes.versions")
 }
 apply<CommonBuildPlugin>()
 
@@ -10,4 +11,10 @@ dependencies {
     implementation(Deps.Kotlin.stdlib)
     api(Deps.AndroidX.Lifecycle.liveDataKtx)
     api(Deps.inject)
+}
+
+tasks.withType<DependencyUpdatesTask> {
+    rejectVersionIf {
+        isNonStable(candidate.version)
+    }
 }
