+++ data/auth/build.gradle.kts
@@ -1,9 +1,10 @@
+import com.github.benmanes.gradle.versions.updates.DependencyUpdatesTask
+
 plugins {
     id("com.android.library")
     kotlin("android")
     kotlin("android.extensions")
     kotlin("kapt")
-    id("com.github.ben-manes.versions")
 }
 apply<CommonBuildPlugin>()
 
@@ -16,4 +17,10 @@ dependencies {
     api(Deps.Dagger.core)
     kapt(Deps.Dagger.compiler)
     implementation(Deps.timber)
+}
+
+tasks.withType<DependencyUpdatesTask> {
+    rejectVersionIf {
+        isNonStable(candidate.version)
+    }
 }
