+++ data/local/build.gradle.kts
@@ -1,9 +1,10 @@
+import com.github.benmanes.gradle.versions.updates.DependencyUpdatesTask
+
 plugins {
     id("com.android.library")
     kotlin("android")
     kotlin("android.extensions")
     kotlin("kapt")
-    id("com.github.ben-manes.versions")
 }
 apply<CommonBuildPlugin>()
 
@@ -15,4 +16,10 @@ dependencies {
     implementation(Deps.AndroidX.coreKtx)
     api(Deps.Dagger.core)
     kapt(Deps.Dagger.compiler)
+}
+
+tasks.withType<DependencyUpdatesTask> {
+    rejectVersionIf {
+        isNonStable(candidate.version)
+    }
 }
