+++ model/build.gradle.kts
@@ -1,11 +1,18 @@
+import com.github.benmanes.gradle.versions.updates.DependencyUpdatesTask
+
 plugins {
     id("com.android.library")
     kotlin("android")
     kotlin("android.extensions")
-    id("com.github.ben-manes.versions")
 }
 apply<CommonBuildPlugin>()
 
 dependencies {
     implementation(Deps.Kotlin.stdlib)
+}
+
+tasks.withType<DependencyUpdatesTask> {
+    rejectVersionIf {
+        isNonStable(candidate.version)
+    }
 }
